+++ packages/components/alert-dialog/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.24](https://github.com/adevinta/spark/compare/@spark-ui/alert-dialog@1.0.23...@spark-ui/alert-dialog@1.0.24) (2024-06-25)
+
+**Note:** Version bump only for package @spark-ui/alert-dialog
+
+## [1.0.23](https://github.com/adevinta/spark/compare/@spark-ui/alert-dialog@1.0.22...@spark-ui/alert-dialog@1.0.23) (2024-06-21)
+
+**Note:** Version bump only for package @spark-ui/alert-dialog
+
+## [1.0.22](https://github.com/adevinta/spark/compare/@spark-ui/alert-dialog@1.0.21...@spark-ui/alert-dialog@1.0.22) (2024-06-20)
+
+**Note:** Version bump only for package @spark-ui/alert-dialog
+
 ## [1.0.21](https://github.com/adevinta/spark/compare/@spark-ui/alert-dialog@1.0.20...@spark-ui/alert-dialog@1.0.21) (2024-06-15)
 
 **Note:** Version bump only for package @spark-ui/alert-dialog