+++ packages/components/icons/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.21.11](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.21.10...@spark-ui/icons@1.21.11) (2024-06-25)
+
+**Note:** Version bump only for package @spark-ui/icons
+
 ## [1.21.10](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.21.9...@spark-ui/icons@1.21.10) (2024-06-06)
 
 **Note:** Version bump only for package @spark-ui/icons