+++ packages/components/radio-group/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.12](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@2.3.11...@spark-ui/radio-group@2.3.12) (2024-06-25)
+
+**Note:** Version bump only for package @spark-ui/radio-group
+
+## [2.3.11](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@2.3.10...@spark-ui/radio-group@2.3.11) (2024-06-20)
+
+**Note:** Version bump only for package @spark-ui/radio-group
+
 ## [2.3.10](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@2.3.9...@spark-ui/radio-group@2.3.10) (2024-06-15)
 
 **Note:** Version bump only for package @spark-ui/radio-group