+++ packages/components/rating/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.16](https://github.com/adevinta/spark/compare/@spark-ui/rating@1.0.15...@spark-ui/rating@1.0.16) (2024-06-25)
+
+**Note:** Version bump only for package @spark-ui/rating
+
+## [1.0.15](https://github.com/adevinta/spark/compare/@spark-ui/rating@1.0.14...@spark-ui/rating@1.0.15) (2024-06-20)
+
+**Note:** Version bump only for package @spark-ui/rating
+
 ## [1.0.14](https://github.com/adevinta/spark/compare/@spark-ui/rating@1.0.13...@spark-ui/rating@1.0.14) (2024-06-15)
 
 **Note:** Version bump only for package @spark-ui/rating