+++ packages/components/tabs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.24](https://github.com/adevinta/spark/compare/@spark-ui/tabs@2.3.23...@spark-ui/tabs@2.3.24) (2024-06-25)
+
+**Note:** Version bump only for package @spark-ui/tabs
+
+## [2.3.23](https://github.com/adevinta/spark/compare/@spark-ui/tabs@2.3.22...@spark-ui/tabs@2.3.23) (2024-06-20)
+
+**Note:** Version bump only for package @spark-ui/tabs
+
 ## [2.3.22](https://github.com/adevinta/spark/compare/@spark-ui/tabs@2.3.21...@spark-ui/tabs@2.3.22) (2024-06-15)
 
 **Note:** Version bump only for package @spark-ui/tabs